Rename "jvm" feature to "jni"
JVM is a platform, it's confusing to use it as a feature.
It should, for example, be possible to build Android (run `make android`)
with the enabled JNI bridge.

I also renamed `jvm.rs` to `jni_bridge.rs` to make its usage clearer.
Note that calling the file `jni.rs` was not possible as then it would
conflict with the `jni` dependency.
